refactor: split dashboard.py into package
Move the 572-line single-file dashboard into a dashboard/ package with
errors, static, routes, server, and per-feature handlers/ submodules.
Public API (serve_dashboard, build_scaler_response, StaticAssets) stays
at taskito.dashboard via re-exports. No behavior changes.
