+++ py_src/taskito/dashboard/handlers/__init__.py
@@ -1,0 +1,1 @@
+"""Route handler functions, organized by feature area."""
