+++ py_src/taskito/dashboard/handlers/logs.py
@@ -1,0 +1,18 @@
+"""Log query route handlers."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from taskito.dashboard.handlers._qs import _parse_int_qs
+
+if TYPE_CHECKING:
+    from taskito.app import Queue
+
+
+def _handle_logs(queue: Queue, qs: dict) -> list:
+    task = qs.get("task", [None])[0]
+    level = qs.get("level", [None])[0]
+    since = _parse_int_qs(qs, "since", 3600)
+    limit = _parse_int_qs(qs, "limit", 100)
+    return queue.query_logs(task_name=task, level=level, since=since, limit=limit)
